Perf

+++ src/ollama.hpp
@@ -14,7 +14,7 @@ struct OllamaClient {
   static OllamaClient from_env_with_model(const std::string& model_override);
 
   bool available() const;
-  std::vector<double> embed(const std::string& text) const;
+  std::vector<float> embed(const std::string& text) const;
 };
 
 } // namespace mnemon
